requirements.txt
Django==4.2.6
django-debug-toolbar==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
Pillow==10.0.1
psycopg2-binary==2.9.7
